Initial Commit
Licensed under the GPLv3 as per previous GeekZone repositories.

After some discussion on Discord we have decided to go with Python and
Django. This should allow fairly easy and rapid development.

Python is a language that many people are familiar with and enjoy
programming in. It easy relatively easy for people to learn and be
productive with.