Allow deployments to test.geek.zone
Resolves #4.

This commit allows members and contributors to deploy their changes to
a testing environment.

The testing environment is currently a temporary Kubernetes cluster
hosted on AWS using their Elastic Kubernetes Service (EKS).

CircleCI builds docker images of the django web app and an nginx proxy
and then pushes them up to the GeekZone Elastic Container
Registry (ECR). CircleCI then applies the kuberenetes configuration
located in the `k8s` folder to the kubernetes cluster.

Docker images are tagged with the circleci workflow id which is unqiue
to every run of the pipeline.

The django service is run using Gunicorn and all requests to it are
proxied via nginx which serves static assets. There might be a better
way of doing this in the long term?