Use PostgreSQL on test.geek.zone
This commit introduces environment variables that allow the database
to be configured at runtime. We store the `DATABASE_URL` in a secret
on the kubernetes cluster.

We are now using `django-environ` with `.env` files to allow easy
config during local development. I have introduced an `.env.example`
file in the repo to make this easier to configure.

The Dockerfiles have also requied an update to allow them to install
the psycopg2 package.

This commit does not include running migrations on the circleci
pipeline, that can be introduced in another commit. At the moment I
have connected my local Django to the postgres database and run the
migrations on there.