Fix registration and login broken
It appeared that login broke after merging issue39. It turns out that
the `clean_password` function was incorrectly dealt with during a
merge conflict and so was returning the entire form data instead of
the plaintext password.

This still allowed someone to register and be auto logged in, but they
were unable to login manually.