Add setup instructions for Ubuntu based Linux
When this project was started we created some generic installation
instructions based on what we thought we would need. These have not
been updated to reflect how we currently setup and develop on the
system.

These instructions are aimed at newer developers who might not be that
familiar with the process of setting up more complex applications that
utilise package managers, migrations, and environment configuration.

This guide does not cover configuring the application using a .env
file, or setting up stripe products as these are operating system
agnostic. When such a guide has been written I will update the linux
guide to point to the configuration guide at the end.