Ensure recaptcha is only loaded if api keys are available
This commit ensures that recaptcha related code is only run/loaded if
the API keys are available.

This awards us two major benefits:
* We will not have to mock out recaptcha in the automated tests. Prior
to this commit registration tests were failing as recaptcha tagged
them as a bot.
* Developers can opt out of running recaptcha locally. This is helpful
because running through a form very quickly can often mark you as a bot
which developers often do.
* This ensures recaptcha is not required to run the system which means
one less thing for people to worry about when they are getting the
repo set up on their machine.

This does of course come with some drawbacks:
* Developers are less likely to implicitly test recaptcha when doing
development as it would most likely be disabled
* There is now extra code in the view and templates to account for the
recaptcha key not being sent.