Log failed and successful payments (#198)
Log failed payments to a `FailedPayment` model and successful payments to
a `Payment` model. When a member pays we also update a `last_payment_time`
on their `Member` model so we can track the last time a member paid for
sand membership. As more membership types are added we will want to expand
on this design - but for now this suits our needs without creating something
complex prematurely.

The intention of the `FailedPayment` is to catch any cases that we have yet
to write code for so we can manually inspect them and sort them out.

The intention of the `Payment` model is simply to track all of the times a
member has successfully sent money to us. The primary use case I imagine for
this is to display this information to members in their dashboard.

I suppose in the future we could also display some kinds of failed payments?

* Added django_extensiosns to requirements.txt

Co-authored-by: Carwyn Nelson <CarwynNelson@users.noreply.github.com>