Make epoch_to_datetime utility timezone aware
It seems that this recently started erroring in our tests, although
I'm not quite sure how.

This patch is taken as a suggestion from
https://stackoverflow.com/questions/18622007/runtimewarning-datetimefield-received-a-naive-datetime

I get the impression that we should make the entire funky_time file
timezone aware, but I have held off doing that until we have further
evidence that doing so is required.