decouple email sending via celery & RabbitMQ
also adds capacity to add other tasks in the future. 

* add rabbit instructions for linux
Includes instructions on how to setup RabbitMQ & Celery on Linux but not Windows as I did not have Windows available.

* add celery
Installs the celery library using requirements.txt and adds them to other areas in the project.

* create email template content, no styling
no styling created as priority at MVP is to get it working. We will add styling later.

* create email content structure
basic email content setup

* add comment to explain usage
one line comment in the code to ensure that variable usage is clear

* create email task
add task_send_email to tasks.py to allow us to send email

* celery basic setup
create basic celery config to just get it working

* add celery cofig, email config
create configuration to activate celery and allow us to send basic emails

* add run celery worker
expand on instructions to get project working

* correct sender and recipient name
I made a typo which used the to name in the to email field. this fixed that.

* add comment to explain to_name use
just comment to ensrue that to_name is clarified

* update default from email to show from name and email
bake in the GZ email address. in the future we might want to make this more flexible but MPV just needs it to work.

* fixup: Format Python code with Black
auto black the code


Co-authored-by: Geek.Zone/Bot <geekzonebot@users.noreply.github.com>