Move StripeWebhook class to fuction
It has become apparent from making the django `has_paid_sand` permission that it
is easier (and more readable IMO) to have all the the stripe webhook data being
handled by functions instead of a class. These functions are place inside of a
new file, payments.py.

Doing so will help with the returning HttpResponses for Stripe so that it knows
we are properly handling payments, whether they are successful or not.