Create separate test cases for membership renewal
Previously the one test case would check wether a member had a renewal_date but
due to the branching there are two different scenarios - a new user is given a
renewal_date 1 year onward from payment, or an existing member is given a new
renewal_date 1 year from the last date.

This is not ideal as if the else statement is deleted / no longer works, but the
if statement continues to work the test will still work, even though the
functionality is broken.