web-632 replace file name check with branch name check
This change takes the simplified approach of only filtering for branch name to see if a CI skip is permitted.

If branch name starts with "doc-" then CI is not required and can branch can be merged, while is branch starts with "junk-" then it will always fail CI, thereby preventing merge.